requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
openai==2.21.0
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.5
beautifulsoup4==4.14.3
openpyxl==3.1.5
pillow==11.3.0
reportlab==4.4.10
python-multipart==0.0.20
google-genai==1.47.0
